+++ sklearn_xarray/tests/test_dataarray_wrappers.py
@@ -6,7 +6,7 @@ import numpy.testing as npt
 from sklearn.base import BaseEstimator, TransformerMixin
 
 from sklearn_xarray import (
-    wrap, TransformerWrapper, ClassifierWrapper, RegressorWrapper)
+    wrap, wrap, wrap, wrap)
 
 
 class DummyEstimator(BaseEstimator):
@@ -78,7 +78,7 @@ def test_dummy_estimator():
 
     y = X
 
-    estimator = RegressorWrapper(DummyEstimator())
+    estimator = wrap(DummyEstimator())
 
     estimator.fit(X)
     estimator.fit(X, X)
@@ -95,7 +95,7 @@ def test_dummy_transformer():
         dims=['sample', 'feature']
     )
 
-    estimator = TransformerWrapper(DummyTransformer())
+    estimator = wrap(DummyTransformer())
 
     estimator.fit(X)
     yp = estimator.transform(X)
@@ -113,7 +113,7 @@ def test_wrapped_transformer():
         dims=['sample', 'feature']
     )
 
-    estimator = TransformerWrapper(StandardScaler()).fit(X)
+    estimator = wrap(StandardScaler()).fit(X)
 
     assert_allclose(X, estimator.inverse_transform(estimator.transform(X)))
 
@@ -126,7 +126,7 @@ def test_ndim_dummy_estimator():
         dims=['sample', 'feat_1', 'feat_2']
     )
 
-    estimator = RegressorWrapper(DummyEstimator())
+    estimator = wrap(DummyEstimator())
 
     estimator.fit(X, X)
     yp = estimator.predict(X)
@@ -146,7 +146,7 @@ def test_reshaping_estimator():
     y = X[:, :2].drop('feature')
     y['dummy'] = y.dummy[:, 0]
 
-    estimator = RegressorWrapper(
+    estimator = wrap(
         ReshapingEstimator(new_shape=(-1, 2)),
         reshapes='feature'
     )
@@ -169,7 +169,7 @@ def test_reshaping_transformer():
     y = X[:, :2].drop('feature')
     y['dummy'] = y.dummy[:, 0]
 
-    estimator = TransformerWrapper(
+    estimator = wrap(
         ReshapingEstimator(new_shape=(-1, 2)),
         reshapes='feature'
     )
@@ -191,7 +191,7 @@ def test_reshaping_estimator_singleton():
 
     y = X[:, 0].drop('feature')
 
-    estimator = ClassifierWrapper(
+    estimator = wrap(
         ReshapingEstimator(new_shape=(-1, 0)),
         reshapes='feature'
     )
@@ -214,7 +214,7 @@ def test_ndim_reshaping_estimator():
     y = X[:, :5, 0].drop(['feat_1', 'feat_2']).rename({'feat_1': 'feature'})
     y['dummy'] = y.dummy[:, 0]
 
-    estimator = RegressorWrapper(
+    estimator = wrap(
         ReshapingEstimator(new_shape=(-1, 5, 0)),
         reshapes={'feature': ['feat_1', 'feat_2']}
     )
@@ -231,9 +231,9 @@ def test_wrap():
     from sklearn.linear_model import LinearRegression
     from sklearn.svm import SVC
 
-    assert isinstance(wrap(StandardScaler()), TransformerWrapper)
-    assert isinstance(wrap(SVC()), ClassifierWrapper)
-    assert isinstance(wrap(LinearRegression()), RegressorWrapper)
+    assert hasattr(wrap(StandardScaler()), 'transform')
+    assert hasattr(wrap(SVC()), 'predict')
+    assert hasattr(wrap(LinearRegression()), 'score')
 
 
 def test_sample_dim():